Use ElectrictyMap instead of electricityMap for identifiers
