+++ src/CarbonAware.DataSources/CarbonAware.DataSources.ElectricityMap/src/CarbonAware.DataSources.ElectricityMap.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\CarbonAware.LocationSources\CarbonAware.LocationSources.Azure\src\CarbonAware.LocationSources.Azure.csproj" />
+    <ProjectReference Include="..\..\..\CarbonAware.Tools\CarbonAware.Tools.ElectricityMapClient\src\CarbonAware.Tools.ElectricityMapClient.csproj" />
+    <ProjectReference Include="..\..\..\CarbonAware\src\CarbonAware.csproj" />
+  </ItemGroup>
+
+</Project>
