+++ src/CarbonAware.DataSources/CarbonAware.DataSources.ElectricityMap/src/InternalsVisibleTo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly:InternalsVisibleTo("CarbonAware.DataSources.ElectricityMap.Tests")]
