+++ src/CarbonAware.DataSources/CarbonAware.DataSources.ElectricityMap/test/LocationConverterTests.cs
@@ -1,0 +1,17 @@
+﻿using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CarbonAware.DataSources.ElectricityMap.Tests
+{
+    public class LocationConverterTests
+    {
+        [SetUp]
+        public void Setup()
+        {
+        }
+    }
+}
