--- src/CarbonAware.DataSources/CarbonAware.DataSources.electricityMap/src/CarbonAware.DataSources.electricityMap.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\CarbonAware.LocationSources\CarbonAware.LocationSources.Azure\src\CarbonAware.LocationSources.Azure.csproj" />
-    <ProjectReference Include="..\..\..\CarbonAware.Tools\CarbonAware.Tools.electricityMapClient\src\CarbonAware.Tools.electricityMapClient.csproj" />
-    <ProjectReference Include="..\..\..\CarbonAware\src\CarbonAware.csproj" />
-  </ItemGroup>
-
-</Project>
