--- src/CarbonAware.DataSources/CarbonAware.DataSources.electricityMap/src/InternalsVisibleTo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly:InternalsVisibleTo("CarbonAware.DataSources.electricityMap.Tests")]
