--- src/CarbonAware.DataSources/CarbonAware.DataSources.electricityMap/test/LocationConverterTests.cs
@@ -1,17 +1,0 @@
-﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CarbonAware.DataSources.electricityMap.Tests
-{
-    public class LocationConverterTests
-    {
-        [SetUp]
-        public void Setup()
-        {
-        }
-    }
-}
