+++ src/CarbonAware.Tools/CarbonAware.Tools.ElectricityMapClient/src/Constants/AuthenticationHeaderTypes.cs
@@ -1,0 +1,8 @@
+﻿namespace CarbonAware.Tools.ElectricityMapClient.Constants;
+
+// TODO: need change for electricityMap Authorization
+internal class AuthenticationHeaderTypes
+{
+    // public const string Basic = "Basic";
+    public const string Token = "auth-token";
+}
