+++ src/CarbonAware.Tools/CarbonAware.Tools.ElectricityMapClient/src/Constants/Paths.cs
@@ -1,0 +1,9 @@
+﻿namespace CarbonAware.Tools.ElectricityMapClient.Constants;
+
+internal class Paths
+{
+    public const string Data = "data";
+    public const string Forecast = "forecast";
+    public const string Login = "login";
+    public const string Historical = "past";
+}
