+++ src/CarbonAware.Tools/CarbonAware.Tools.ElectricityMapClient/src/Constants/QueryStrings.cs
@@ -1,0 +1,11 @@
+﻿namespace CarbonAware.Tools.ElectricityMapClient.Constants;
+
+internal class QueryStrings
+{
+    public const string countryCodeAbbreviation = "countryCode";
+    public const string StartTime = "starttime";
+    public const string EndTime = "endtime";
+    public const string Latitude = "latitude";
+    public const string Longitude = "longitude";
+    public const string Username = "username";
+}
