+++ src/CarbonAware.Tools/CarbonAware.Tools.ElectricityMapClient/src/ElectricityMapClientException.cs
@@ -1,0 +1,11 @@
+﻿using CarbonAware.Interfaces;
+using System;
+
+namespace CarbonAware.Tools.ElectricityMapClient;
+
+public class ElectricityMapClientException : Exception
+{
+    public ElectricityMapClientException(string message) : base(message)
+    {
+    }
+}
