+++ src/CarbonAware.Tools/CarbonAware.Tools.ElectricityMapClient/src/Model/LoginResult.cs
@@ -1,0 +1,16 @@
+﻿using System.Text.Json.Serialization;
+
+namespace CarbonAware.Tools.ElectricityMapClient.Model;
+
+/// <summary>
+/// Serializable object describing the electricityMap login response object.
+/// </summary>
+[Serializable]
+public record LoginResult
+{
+    /// <summary>
+    /// The Bearer token used to authenticate future requests.
+    /// </summary>
+    [JsonPropertyName("token")]
+    public string Token { get; set; } = string.Empty;
+}
