+++ src/CarbonAware.Tools/CarbonAware.Tools.ElectricityMapClient/test/AssemblyInfo.cs
