--- src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Configuration/ConfigurationException.cs
@@ -1,11 +1,0 @@
-﻿namespace CarbonAware.Tools.electricityMapClient.Configuration;
-
-/// <summary>
-/// An exception class thrown when the electricityMap client is misconfigured.
-/// </summary>
-public class ConfigurationException : Exception
-{
-    public ConfigurationException(string message) : base(message)
-    {
-    }
-}
