--- src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Constants/AuthenticationHeaderTypes.cs
@@ -1,8 +1,0 @@
-﻿namespace CarbonAware.Tools.electricityMapClient.Constants;
-
-// TODO: need change for electricityMap Authorization
-internal class AuthenticationHeaderTypes
-{
-    // public const string Basic = "Basic";
-    public const string Token = "auth-token";
-}
