--- src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Constants/Paths.cs
@@ -1,9 +1,0 @@
-﻿namespace CarbonAware.Tools.electricityMapClient.Constants;
-
-internal class Paths
-{
-    public const string Data = "data";
-    public const string Forecast = "forecast";
-    public const string Login = "login";
-    public const string Historical = "past";
-}
