--- src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/Constants/QueryStrings.cs
@@ -1,11 +1,0 @@
-﻿namespace CarbonAware.Tools.electricityMapClient.Constants;
-
-internal class QueryStrings
-{
-    public const string countryCodeAbbreviation = "countryCode";
-    public const string StartTime = "starttime";
-    public const string EndTime = "endtime";
-    public const string Latitude = "latitude";
-    public const string Longitude = "longitude";
-    public const string Username = "username";
-}
