--- src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/InternalsVisibleTo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly:InternalsVisibleTo("CarbonAware.Tools.electricityMapClient.Tests")]
