--- src/CarbonAware.Tools/CarbonAware.Tools.electricityMapClient/src/electricityMapClienttException.cs
@@ -1,11 +1,0 @@
-﻿using CarbonAware.Interfaces;
-using System;
-
-namespace CarbonAware.Tools.electricityMapClient;
-
-public class electricityMapClientException : Exception
-{
-    public electricityMapClientException(string message) : base(message)
-    {
-    }
-}
